Loader Script

+++ Assets/Scripts/Kepler/GravitySource.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public class GravitySource
 {
     public Transform AttractorObject;
-    public float AttractorMass = 1000f;
-    public float GravityConstant = 0.1f;
+    public double AttractorMass = 1000f;
+    public double GravityConstant = 0.1f;
 }
